+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 - [Feature] Add purple **Badge** color option ([#1377](https://github.com/optimizely/oui/pull/1377))
 - [Feature] Added new Storybook examples detailing spacing helper classes ([#1378](https://github.com/optimizely/oui/pull/1378))
 - [Feature] Add new **Form** component and updated the **Fieldset** component to be accessible ([#1374](https://github.com/optimizely/oui/pull/1374))
+- [Feature] Add `leftIcon` and `rightIcon` to **Button** ([#1383](https://github.com/optimizely/oui/pull/1383))
+- [Patch] Fix console errors for **SearchPicker** ([#1385](https://github.com/optimizely/oui/pull/1385))
 
 ## 46.10.0 - 2020-07-14